+++ b2sdk/_v3/__init__.py
@@ -127,6 +127,7 @@ from b2sdk.stream import WritingStreamWithProgress
 
 from b2sdk.transfer.inbound.downloaded_file import DownloadedFile
 from b2sdk.transfer.inbound.downloaded_file import MtimeUpdatedFile
+from b2sdk.transfer.inbound.download_manager import DownloadManager
 
 from b2sdk.transfer.outbound.outbound_source import OutboundTransferSource
 from b2sdk.transfer.outbound.copy_source import CopySource
@@ -136,6 +137,7 @@ from b2sdk.transfer.outbound.upload_source import UploadSourceLocalFile
 from b2sdk.transfer.outbound.upload_source import UploadSourceLocalFileRange
 from b2sdk.transfer.outbound.upload_source import UploadSourceStream
 from b2sdk.transfer.outbound.upload_source import UploadSourceStreamRange
+from b2sdk.transfer.outbound.upload_manager import UploadManager
 
 from b2sdk.transfer.emerge.planner.upload_subpart import CachedBytesStreamOpener
 from b2sdk.transfer.emerge.write_intent import WriteIntent
@@ -144,9 +146,6 @@ from b2sdk.transfer.emerge.write_intent import WriteIntent
 
 from b2sdk.transfer.inbound.downloader.abstract import AbstractDownloader
 from b2sdk.transfer.outbound.large_file_upload_state import LargeFileUploadState
-from b2sdk.transfer.inbound.downloader.parallel import AbstractDownloaderThread
-from b2sdk.transfer.inbound.downloader.parallel import FirstPartDownloaderThread
-from b2sdk.transfer.inbound.downloader.parallel import NonHashingDownloaderThread
 from b2sdk.transfer.inbound.downloader.parallel import ParallelDownloader
 from b2sdk.transfer.inbound.downloader.parallel import PartToDownload
 from b2sdk.transfer.inbound.downloader.parallel import WriterThread
@@ -215,3 +214,4 @@ from b2sdk.cache import DummyCache
 from b2sdk.cache import InMemoryCache
 from b2sdk.http_constants import SRC_LAST_MODIFIED_MILLIS
 from b2sdk.session import B2Session
+from b2sdk.utils.thread_pool import ThreadPoolMixin