+++ b2sdk/transfer/outbound/copy_manager.py
@@ -8,7 +8,6 @@
 #
 ######################################################################
 
-import concurrent.futures as futures
 import logging
 from typing import Optional
 
@@ -17,52 +16,22 @@ from b2sdk.http_constants import SSE_C_KEY_ID_FILE_INFO_KEY_NAME
 from b2sdk.exception import AlreadyFailed, CopyArgumentsMismatch, SSECKeyIdMismatchInCopy
 from b2sdk.file_lock import FileRetentionSetting, LegalHold
 from b2sdk.raw_api import MetadataDirectiveMode
-from b2sdk.utils import B2TraceMetaAbstract
+from b2sdk.transfer.transfer_manager import TransferManager
+from b2sdk.utils.thread_pool import ThreadPoolMixin
 
 logger = logging.getLogger(__name__)
 
 
-class CopyManager(metaclass=B2TraceMetaAbstract):
+class CopyManager(TransferManager, ThreadPoolMixin):
     """
     Handle complex actions around server side copy to free raw_api from that responsibility.
     """
 
     MAX_LARGE_FILE_SIZE = 10 * 1000 * 1000 * 1000 * 1000  # 10 TB
 
-    def __init__(self, services, max_copy_workers=10):
-        """
-        :param b2sdk.v2.Services services:
-        :param int max_copy_workers: a number of copy threads
-        """
-        self.services = services
-
-        self.copy_executor = None
-        self.max_workers = max_copy_workers
-
     @property
     def account_info(self):
         return self.services.session.account_info
-
-    def set_thread_pool_size(self, max_workers):
-        """
-        Set the size of the thread pool to use for uploads and downloads.
-
-        Must be called before any work starts, or the thread pool will get
-        the default size.
-
-        :param int max_workers: maximum allowed number of workers in a pool
-        """
-        if self.copy_executor is not None:
-            raise Exception('thread pool already created')
-        self.max_workers = max_workers
-
-    def get_thread_pool(self):
-        """
-        Return the thread pool executor to use for uploads and downloads.
-        """
-        if self.copy_executor is None:
-            self.copy_executor = futures.ThreadPoolExecutor(max_workers=self.max_workers)
-        return self.copy_executor
 
     def copy_file(
         self,
@@ -79,7 +48,7 @@ class CopyManager(metaclass=B2TraceMetaAbstract):
     ):
         # Run small copies in the same thread pool as large file copies,
         # so that they share resources during a sync.
-        return self.get_thread_pool().submit(
+        return self._thread_pool.submit(
             self._copy_small_file,
             copy_source,
             file_name,
@@ -103,7 +72,7 @@ class CopyManager(metaclass=B2TraceMetaAbstract):
         destination_encryption: Optional[EncryptionSetting] = None,
         source_encryption: Optional[EncryptionSetting] = None,
     ):
-        return self.get_thread_pool().submit(
+        return self._thread_pool.submit(
             self._copy_part,
             large_file_id,
             part_copy_source,